requirements.txt
sqlalchemy>=1.4.0
pydantic>=1.8.0
PyYAML>=6.0
aiohttp>=3.8.0
Pillow>=8.0.0
cachetools>=4.2.0
imagehash>=4.2.0
nonebot2>=2.0.0
nonebug>=0.2.0
nonebot-adapter-onebot>=2.0.0
nonebot-plugin-datastore>=0.1.0
nonebot-plugin-txt2img>=0.3.0
nonebot-plugin-apscheduler>=0.2.0
nonebot-plugin-userinfo>=0.1.0
nonebot-plugin-localstore>=0.1.0
nonebot-plugin-send-anything-anywhere = "^0.1.0"
pytest>=6.2.0
pytest-asyncio>=0.14.0
pytest-mock>=3.5.0
